* Fix gitlab-runner configuration for tt and snow.

* Use Ninja for some CI/regress builds.

* Begin supporting cce-13.0.1.

* NN Prediction

* Debug regression failures.

* Debug regression failures.

* Debug regression failures.

* Fix regression runs that use Ninja.

* Fix bad merge

* Fix bad merge

* More regression fixes.

* More regression script fixes.

* NN cleanup

* Another try to get MAKEFLAGS right.

* Switch gcov build back to Makefiles.

* Starting using LAP SE 2.0

* Upgrade to LAP SE-2.0 gcc for CI

* Try to fix downstream-deploy target.

* Create regression/deploy targets for ccsnet gcc831, gcc1120, llvm12

* Register new env names with CI system.

* Use CPP to get the OpenMP syntax correct for different compilers/platforms.

* Add a deployment job for ccscs1.

* Add a deployment job for ccscs9.

* Add deployment jobs to support capsaicin on ccs-net.

* Updates for ccs-net, gcc 8.3.1 -> 8.5.0

* Make sanity check for available modules more robust.

* fix typo.

* fix typo.

* fix typo.

* Add the llvm12-release deploy build on ccsnet

* Add long long specialization of global_min to rtt_c4

* Increment default environments for ccs-net and tt.

* Can we remove SP.hh?

* ccsnet: Adopt new environments for CI.

* clang-tidy deploy needs llvm12.

* Use new scheduler parameters for snow.

* Fix the ccs1_deb_gcc_1120_deploy target.

* Add new valgrind suppressions for openmpi@4.1.1

* Rework code to address build warnings from oneapi.

* Why does valgrind CI pass when valgrind regress fails?

* Add version check for flake8 to git pre-commit install hooks

* Use std functions instead of draco wrappers for isinf, etc.

* Suppress a OneApi build warning.

* Memcheck and autodoc fixes.

* Change fp-model to precise.

* Improve robustness of covrep target.

* Create a check_style CI that runs on Darwin.

* Fix a bad line wrap.

* Attempt to create a rule to block merge if coverage decreases.

* Restrict cell faces to be triangles for 3D random python mesh type.

* Grizzly is gone.

* Updates to support MSVC 2022.

* Fix KDE index in small windows and add more coarse index information

* Start running MSVC regression via gitlab runner scheduled job.

* Fix deployment rule for MSVC.

* Fix typo.

* For MSVC nightly add PATHs for numdiff and gfortran.

* Bug fixes for MSVC nightly regression.

* Fix sitename and buildname for msvc regressions.

* Cleanup MSVC CI scripts.

* Tweak warning flags for OneAPI.

* For MSVC gitlab-runner jobs, place the build in a CMAKE_BUILD_TYPE subdirectory.

* MSVC gitlab-runners: Shorten paths.

* Improve interoperability with vcpkg.

* One more fix to eliminate MSVC's defaultlib warning.

* MSVC 2022: Re-enable openmp via /openmp:experimental build option.

* Add support for building with SE 2.2.0 (cce-14) on rznevada.

* Retire lapack-wrap.

* Weighted kde

* Fix typo in bash function findsymbol().

* Try to allow manual deployment option.

* Use new envs (openmpi@4.1.1) on darwin.

* Remove orte_tmpdir_base option from MPI launch command.

* Merge draco-7_14_0 back to develop

* Add a script to find VTune profiling API libraries

* Cleanup code that does vendor discovery.

* Merge draco 7_14_1 release back to develop

* Disable tt/cce CI pipelines for now

* Minor tweaks to trigger CI testing for tt.

* Update Changelog for patch release

* Eliminate qt dependency.

* Don't cache DRACO_DBC_LEVEL for client projects.

* Update rules for copying dll files to lib/exe directory.

* Make cmake@3.21.0 the minimum required version

* Fix issues reported by latest OneAPI

* Add the 'batch' CI tag as requested by admins.

* Change default dev env to LAP SE 2.2(.0)

* Standardize OpenMP_FOUND variable names.

* Remove release-related scripts (but hopefully not scripts used elsewhere in draco)

* Tweak some code for cleaner compile on p9 with xl@16, gcc@9.3.0, cuda@11.7.0

* Move ccs-net CI back to original machines

* Change (valid) code that causes linking fits for XL16.1+gcc9.3

* Updates to allow clang-cl builds under MSVC.

* Fix clang-tidy warning.

* Fix srun options for rzvernal.

* Add support for nvhpc compilers.

* Enable clang-tidy hicpp-* checks for ds++.

* Change logic for cce/13 environment on tt.

* Apply clang-tidy hicpp changes to c4 and cdi.

* Updates for CI/Regression (stop valgrind; add SPR/oneapi)

* More updates recommended by clang-tidy hicpp-*

* Break CSK dependency in draco by providing needed interpolation/data read functions, testing.

* Make 1d interpolation gpu-friendly

* Fix assignment/move operations for UnitSystem and UnitSystemType

* Remove noexcept for UnitSystemType move assignment operator

* More clang-tidy hicpp- updates.

* Another round of clang-tidy hicpp- fixes.

* Minor fixes after enabling hicpp- checks for clang-tidy.

* Improve test coverage.

* Remove unused RNG files and associated tests

* Prevent multiple conditions on a cell face in Draco_Mesh.

* Updates to address static analysis warnings.

* Loosen tolerance to allow test to pass for nvc++

* Remove reference to deleted compton_interface dir

* Clean up some yaml formatting.

* Cleanup some documentation.

* Fix new badges displayed in `README.md`

* Fixes to accommodate the new version of super-linter.

* Retire oneapi CI/regress for tt.

* Add manual deploy rule for vs22.

* Promote toggle-compiler-flags to be a cmake function.

* Add a check that fails if OpenMP verion is too old.

* Tweak the toggle_compiler_flags function to fix client build issues with CAFS.

* Add functions to get/set OMP schedule.

* Prototype new cmake presets capability.

* Demo lapse/2.2-intel env.

* Retire intel@19 pipelines for trinitite.

* Fix EXTRA_CMAKE_ARGS in yaml files.

* Update default dev-env for tt.

* New xl env on power9 darwin

* Fix deploy dir for darwin.

* Avoid CAFS logic when not MSVC

* Corrected syntax error

* Add support for HIP

* Post MR-173 fixes

* Remove now-unused threading support wrappers from draco; add an explicit toggle for OMP support

* Add double quotes to prevent configure failures when toggling potentially empty OpenMP variables

* Update change long and release info for draco-7_15_0

* remove github workflows

* remove github appveyor.yml

---------

+++ src/ds++/dbc.hh
@@ -4,7 +4,7 @@
  * \author Kent G. Budge
  * \date   Wed Jan 22 15:18:23 MST 2003
  * \brief  Extensions to the STL algorithm library
- * \note   Copyright (C) 2016-2021 Triad National Security, LLC., All rights reserved.
+ * \note   Copyright (C) 2010-2022 Triad National Security, LLC., All rights reserved.
  *
  * This header defines several function templates that perform common numerical operations not
  * standardized in the STL algorithm header. It also defines some useful STL-style predicates. These
@@ -14,6 +14,8 @@
 
 #ifndef rtt_dsxx_dbc_hh
 #define rtt_dsxx_dbc_hh
+
+#include <cstddef>
 
 namespace rtt_dsxx {
 
@@ -31,7 +33,7 @@ bool is_strict_monotonic_decreasing(Forward_Iterator first, Forward_Iterator con
 
 //! Check whether a matrix is symmetric.
 template <typename Random_Access_Container>
-bool is_symmetric_matrix(Random_Access_Container const &A, unsigned const n,
+bool is_symmetric_matrix(Random_Access_Container const &A, size_t const n,
                          double const tolerance = 1.0e-12);
 
 //! Return the positive difference of the arguments.
@@ -44,6 +46,7 @@ inline Ordered_Group_Element dim(Ordered_Group_Element a, Ordered_Group_Element 
 #include "dbc.i.hh"
 
 #endif // rtt_dsxx_dbc_hh
+
 //------------------------------------------------------------------------------------------------//
 // end of dbc.hh
 //------------------------------------------------------------------------------------------------//
